photos moved to folder "photos"
