+++ react-chat/src/pages/PageChatList/PageChatList.jsx
@@ -4,10 +4,10 @@ import SearchIcon from '@mui/icons-material/Search';
 import EditIcon from '@mui/icons-material/Edit';
 //import AttachmentIcon from '@mui/icons-material/Attachment';
 import './PageChatList.scss';
-import mycat from '../../mycat.jpg';
-import polina from '../../polina.jpg';
-import photonics from '../../photonics.jpg';
-import kittens from '../../kittens.jpg';
+import mycat from '../../photos/mycat.jpg';
+import polina from '../../photos/polina.jpg';
+import photonics from '../../photos/photonics.jpg';
+import kittens from '../../photos/kittens.jpg';
 import {Button} from '../../components';
 import {ChatHead} from '../../components';
 import {Chat} from '../../components';
